changed app.py to main.py, added new flask file named app.py and index.html
