+++ proposture/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+  <head>
+      <title>ProPosture App</title>
+  </head>
+  <body>
+      <h1>Welcome to ProPosture</h1>
+
+      <div>
+          <video id="videoPlayer" controls autoplay>
+              <source src="{{ video_url }}" type="video/mp4">
+          </video>
+      </div>
+
+      <p>This is the main page of the ProPosture web application.</p>
+  </body>
+</html>
